+++ src/include/config/constants.py
@@ -30,7 +30,7 @@ from pathlib import Path
 
 from include.config.version import Version
 
-CORE_VERSION = Version("0.4.1.260811_alpha")
+CORE_VERSION = Version("0.5.0.260812_alpha")
 PROTOCOL_VERSION = 21
 
 ROOT_ABSPATH = Path(__file__).resolve().parents[2]